refactor!: add runtime + types folders
