+++ scripts/getPositions.ts
@@ -2,8 +2,8 @@
 /* eslint-disable no-console */
 import yargs from 'yargs'
 import BigNumber from 'bignumber.js'
-import { Token } from '../src/positions'
-import { getPositions } from '../src/getPositions'
+import { Token } from '../src/types/positions'
+import { getPositions } from '../src/runtime/getPositions'
 
 const argv = yargs(process.argv.slice(2))
   .usage('Usage: $0 --address <address>')
