+++ scripts/getShortcuts.ts
@@ -1,7 +1,7 @@
 // Helper script to get the shortcuts
 /* eslint-disable no-console */
 import yargs from 'yargs'
-import { getShortcuts } from '../src/getShortcuts'
+import { getShortcuts } from '../src/runtime/getShortcuts'
 
 const argv = yargs(process.argv.slice(2))
   .usage('Usage: $0 --apps app1[,app2]')
