--- src/data.json
@@ -1,4 +1,0 @@
-{
-  "some": "data",
-  "important": true
-}
