--- src/index.test.ts
@@ -1,7 +1,0 @@
-import { main } from './index'
-
-describe(main, () => {
-  it("should return 'Hello, world!'", () => {
-    expect(main()).toBe('Hello, world!')
-  })
-})
