--- src/index.ts
@@ -1,3 +1,0 @@
-export function main() {
-  return 'Hello, world!'
-}
