+++ src/runtime/getShortcuts.e2e.ts
