+++ src/runtime/getShortcuts.ts
