+++ src/types/shortcuts.ts
